Copy licenses into container image
This adds bottlerocket-license-scanner's license scanning output and the
repository's LICENSEs in the built container image. The clarify.toml has
been added to clear up the unknown license in the sigs.k8s.io/yaml
dependency.

By using the scanner, the build process now relies on the bottlerocket
SDK container and is downloaded automatically for builds.