fix agent stuck on drain bug
Some customers are using StatefulSet which makes pod's name unchangeable,
and we use the name of the evicted pod and query the k8s API on wait_for_deletion
fucntion, waiting for it to 404 before declaring a pod deleted.
Therefore, current logic can't handle this sepcial case. We add a new if
statement condition to compare two pods by uid to make sure if they are the
same pod. If they are same pod, brupop should wait until it be deleted,
otherwise we should recognize the old pod has already been deleted and
re-lives on other node.