Make sure cron takes precendence over time schedule
Existing code would error out if both a cron expression and a time
window were provided. This changes the behavior to always give cron the
priority and ignore any time window settings if also provided.

Also updates and reorders the README documentation on setting the
schedule so the current and preferred cron information comes first,
followed by the deprecated time window information. A note is also added
to the time window portion calling out the cron expressions will be used
if both are provided.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>